Move ArmFeedForward to utils
